requirements.txt
streamlit
google-generativeai
gtts
textstat
python-dotenv
pandas